% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_VaR.R
\name{calculate_VaR}
\alias{calculate_VaR}
\title{Calculate Value at Risk (VaR) and Expected Shortfall (ES) of a portfolio}
\usage{
calculate_VaR(
  portfolio_returns,
  alpha = 0.05,
  method = "historical",
  distribution = NULL
)
}
\arguments{
\item{portfolio_returns}{A numeric vector of portfolio returns.}

\item{alpha}{A numeric value between 0 and 1 representing the significance level.}

\item{method}{A character string specifying the method to be used ('historical', 'gaussian', 't-student' or 'ewma'). Default is 'historical'.}

\item{distribution}{If method = 'ewma', then a character string specifying which distribution to use for calculating VaR ('gaussian' or 't-student').}
}
\value{
A list containing the VaR, ES, Realized Exceedances and Expected Exceedances.
}
\description{
This function calculates the Value at Risk (VaR) and Expected Shortfall (ES) using
one of the three methods: historical, gaussian (parametric), and t-student.
}
\examples{
# Example with historical method
returns <- rnorm(1000, mean=0.0005, sd=0.01)
calculate_VaR(returns, alpha=0.05, method="historical")

# Example with gaussian method
calculate_VaR(returns, alpha=0.05, method="gaussian")

# Example with t-student method
calculate_VaR(returns, alpha=0.05, method="t-student")
}
