% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_VaR_EWMA.R
\name{calculate_VaR_EWMA}
\alias{calculate_VaR_EWMA}
\title{Calculate VaR using Exponentially Weighted Moving Average (EWMA) method}
\usage{
calculate_VaR_EWMA(portfolio_returns, alpha = 0.05, distribution = "gaussian")
}
\arguments{
\item{portfolio_returns}{A numeric vector of portfolio returns.}

\item{alpha}{A numeric value between 0 and 1, representing the significance level.}

\item{distribution}{A character string specifying the distribution ('gaussian' or 't-student').}
}
\value{
Returns a numeric value representing the VaR.
}
\description{
This function calculates Value at Risk (VaR) of portfolio returns using the EWMA method for
estimating volatility, with an option to assume normal or Student's t-distribution.
}
\examples{
\dontrun{
portfolio_returns <- rnorm(1000)

# Calculate VaR using normal distribution
VaR_EWMA_normal <- calculate_EWMA_VaR(portfolio_returns, alpha = 0.05, distribution = "gaussian")

# Calculate VaR using t-distribution
VaR_EWMA_t <- calculate_EWMA_VaR(portfolio_returns, alpha = 0.05, distribution = "t-student")
}
}
