% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_VaR_monte_carlo.R
\name{scale_VaR_monte_carlo}
\alias{scale_VaR_monte_carlo}
\title{Calculate Value-at-Risk (VaR) for different time horizons using Monte Carlo methods.}
\usage{
scale_VaR_monte_carlo(H = 5, N = 10000, alpha = 0.05, method, params = NULL)
}
\arguments{
\item{H}{An integer specifying the maximal time horizon. Default is 5.}

\item{N}{An integer specifying the number of simulated paths for each time horizon. Default is 1e4.}

\item{alpha}{A numeric value between 0 and 1 specifying the significance level for the VaR. Default is 0.05.}

\item{method}{A character string specifying the method for generating returns. Currently, only 'gaussian' 
is supported, which means returns are assumed to follow a normal distribution.}

\item{params}{A list of additional parameters. If method = guassian}
}
\value{
A numeric vector of length H containing the VaR at each time horizon, then params must contain mean and sd.
}
\description{
This function calculates the Value-at-Risk (VaR) at different time horizons using a Monte Carlo simulation. 
Returns generating process is provided by the user with additional parameters. It calculates the VaR at each time horizon 
by generating N different return paths of h days, and then taking the cumulative sum of returns over each path.
It then finds the alpha-quantile of these cumulative returns, which represents the VaR at the given significance level 
and time horizon.
}
\examples{
# To calculate the VaR at 5 different horizons assuming zero mean and unit standard deviation returns:
scale_VaR_monte_carlo(H = 5, N = 1e4, alpha = 0.05, method = 'gaussian', params = list(mean = 0, sd = 1))
}
